move flow route to index
