+++ fresh.gen.ts
@@ -2,24 +2,24 @@
 // This file SHOULD be checked into source version control.
 // This file is automatically updated during development when running `dev.ts`.
 
-import * as $0 from "./routes/[name].tsx";
-import * as $1 from "./routes/api/joke.ts";
-import * as $2 from "./routes/flow/[info].tsx";
-import * as $3 from "./routes/flow/index.tsx";
-import * as $4 from "./routes/grpc.tsx";
-import * as $5 from "./routes/index.tsx";
-import * as $6 from "./routes/login.tsx";
+import * as $0 from "./routes/[info].tsx";
+import * as $1 from "./routes/[name].tsx";
+import * as $2 from "./routes/api/joke.ts";
+import * as $3 from "./routes/grpc.tsx";
+import * as $4 from "./routes/index.tsx";
+import * as $5 from "./routes/login.tsx";
+import * as $6 from "./routes/rules.tsx";
 import * as $$0 from "./islands/flow.tsx";
 
 const manifest = {
   routes: {
-    "./routes/[name].tsx": $0,
-    "./routes/api/joke.ts": $1,
-    "./routes/flow/[info].tsx": $2,
-    "./routes/flow/index.tsx": $3,
-    "./routes/grpc.tsx": $4,
-    "./routes/index.tsx": $5,
-    "./routes/login.tsx": $6,
+    "./routes/[info].tsx": $0,
+    "./routes/[name].tsx": $1,
+    "./routes/api/joke.ts": $2,
+    "./routes/grpc.tsx": $3,
+    "./routes/index.tsx": $4,
+    "./routes/login.tsx": $5,
+    "./routes/rules.tsx": $6,
   },
   islands: {
     "./islands/flow.tsx": $$0,
